Fixed all errors up to this point! - Added a default constructor to the Satellite and Ground Station objects, Added a << operator to Structure (can overload this in Satellite and Ground Station in the future), and changed the weight in Vertex's add_edge function from int to double

+++ include/Vertex.h
@@ -64,7 +64,7 @@ class Vertex{
 		}
 
 		// Adding an Edge to a Vertex
-		void add_edge( unsigned int destin, int weight ){
+		void add_edge( unsigned int destin, double weight ){
 
 			// If not, put the edge on the back of the array
 			edges.push_back( Edge(destin, weight) );
